ml/pipelines/father_death_predictor/requirements.txt
pyswisseph>=2.10.3
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
lightgbm>=4.0
optuna>=3.4
shap>=0.43
lifelines>=0.27
matplotlib>=3.7
seaborn>=0.12
pyarrow>=14.0
scipy>=1.10
pyyaml>=6.0
